feat: implement knowledge editor UI and modularize interaction logic for page management
